Add spec for preventing overriding Spaces with non Blank moves.
Once a Space is “claimed” via a Move, the Space’s Move should not be
mutable. I’m not sure if this is possible to express through a type, so
for now I plan to add a guard clause. If the guard clause is true, we
can confidently make the move.