Minimax v2
The big mistake I was making in my previous minimax version was not
associating scores to the root positions. Instead, I was only able to
get scores and moves for the moves that were terminal. Because this was
causing the root position moves (where ply == 0) to be lost, incorrect
moves were selected based on the terminal move, which is too far into
the future to be effective from the perspective of the “current” state
of the game board.